backend/requirements.txt
Django>=4.0
gunicorn>=20.0.4
mysqlclient>=2.0
django-allauth==64.1.0
requests==2.32.3
PyJWT==2.9.0
cryptography==43.0.0
Pillow>=8.0
djangorestframework==3.15.2
django-filter==24.3
# Markdown==3.7
django-cors-headers==4.5.0
dj-rest-auth==6.0.0
djangorestframework-simplejwt==5.3.1
# djoser==2.2.3
django-environ==0.11.2
python-decouple>=3.8
